remove direct connect browser from `react` pkg -- use ws
